Added Global Exception Handler

+++ backend/src/main/java/com/project_manager/TrackFlow/service/impl/InvitationServiceImpl.java
@@ -1,10 +1,10 @@
 package com.project_manager.TrackFlow.service.impl;
 
+import com.project_manager.TrackFlow.Exceptions.InvalidRequest;
 import com.project_manager.TrackFlow.model.Invitation;
 import com.project_manager.TrackFlow.repository.InvitationRepository;
 import com.project_manager.TrackFlow.service.EmailService;
 import com.project_manager.TrackFlow.service.InvitationService;
-import jakarta.mail.MessagingException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,7 +21,7 @@ public class InvitationServiceImpl implements InvitationService {
     private EmailService emailService;
 
     @Override
-    public void sendInvitation(String email, Integer projectId) throws MessagingException {
+    public void sendInvitation(String email, Integer projectId) {
         String inviteToken = UUID.randomUUID().toString();
         Invitation invitation = new Invitation();
         invitation.setEmail(email);
@@ -34,19 +34,19 @@ public class InvitationServiceImpl implements InvitationService {
     }
 
     @Override
-    public Invitation acceptInvitation(String token, Integer userId) throws Exception {
+    public Invitation acceptInvitation(String token, Integer userId) {
         Optional<Invitation> invitation = invitationRepo.findByToken(token);
         if(invitation.isEmpty()){
-            throw new Exception("Invalid Invitation token");
+            throw new InvalidRequest("Invalid Invitation token");
         }
         return invitation.get();
     }
 
     @Override
-    public String getTokenByUserMail(String userEmail) throws Exception{
+    public String getTokenByUserMail(String userEmail) {
         Optional<Invitation> invitation = invitationRepo.findByEmail(userEmail);
         if(invitation.isEmpty()){
-            throw new Exception("Token not Found");
+            throw new InvalidRequest("Token not Found");
         }
         return invitation.get().getToken();
     }
@@ -54,6 +54,9 @@ public class InvitationServiceImpl implements InvitationService {
     @Override
     public void deleteToken(String token) {
         Optional<Invitation> invitation = invitationRepo.findByToken(token);
+        if(invitation.isEmpty()){
+            throw new InvalidRequest("Token not Found");
+        }
         invitationRepo.delete(invitation.get());
     }
 }
